+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2023.08.05 v5.4.0
 
 1. new `shuffle` `randomNumbers` function, see: [shuffle](https://github.com/saqqdy/js-cool#shuffle), [randomNumbers](https://github.com/saqqdy/js-cool#randomnumbers)
-2. band new randomString function, see: [randomString](https://github.com/saqqdy/js-cool#randomstring)
+2. Band new randomString function, see: [randomString](https://github.com/saqqdy/js-cool#randomstring)
 
 ## 2023.08.02 v5.3.2